FUMA-107 FUMA blade templates clean-up.
Move head, nav, footer, and javascript to the partial folder.
Using sections for stylesheets and scripts in common as well as for stylesheets and scripts for a specific view.
Get rid of multiple instances of the same stylesheets and scripts.
Solve some javascript errors mostly on the registration/login views.