Dispatching jobs after all open database ops have been committed.
In the database, some jobs have the QUEUE status even though they were already run.

https://github.com/vufuma/FUMA-webapp/pull/16#issuecomment-1587031019

Brief description of the problem: 

The workflow of a job is as follows:

1. Job creation (marked as new in the db) (this is so fast that we dont even see it).

2. Job is marked as "QUEUED" and the IMEDIATELY

3. Job is dispatched in the queue.

4. When there is an available worker, they pick up this job asd start execution

5. The first step of the execution is to mark the job as RUNNING

Now although this algorithm seems deterministic, for some reasons the 2 step (since it requires DB interactions) is getting delayed by the db. This can happen if the db's load is very high. Then the job is dispatched and the worker starts executing marking it as "RUNNING". At some point tho, the 2 step executes as well marking the job as QUEUED and although the job is still running, or is finished the last db interaction is marking it as QUEUED which is what we see in the logs.